+++ Cargo.toml
@@ -1,3 +1,2 @@
 [workspace]
-members = ["aj", "aj_core", "aj_macro"]
-exclude = ["examples/**"]
+members = ["aj", "aj_core", "aj_macro", "examples/*"]
