+++ aj_core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "aj_core"
-version = "0.6.6"
+version = "0.7.0-a.1"
 edition = "2021"
 authors = ["cptrodgers <me@cptrodgers.com>"]
 description = "Background Job based on Actix"
