+++ examples/in-memory/src/print_job.rs
@@ -1,0 +1,31 @@
+use aj::{
+    async_trait,
+    export::core::serde::{Deserialize, Serialize},
+    BackgroundJob, Executable, JobContext,
+};
+use chrono::{DateTime, Utc};
+
+pub fn get_now() -> DateTime<Utc> {
+    Utc::now()
+}
+
+#[derive(BackgroundJob, Serialize, Deserialize, Debug, Clone)]
+pub struct Print {
+    pub number: i32,
+}
+
+#[async_trait]
+impl Executable for Print {
+    type Output = Result<(), ()>;
+
+    async fn execute(&mut self, _context: &JobContext) -> Self::Output {
+        println!("Hello Job {}, {}", self.number, get_now());
+        Err(())
+    }
+
+    // Determine where your job is failed.
+    // For example, check job output is return Err type
+    async fn is_failed_output(&self, job_output: &Self::Output) -> bool {
+        job_output.is_err()
+    }
+}
