+++ examples/redis_backend/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-aj = { version = "=0.6.4" }
+aj = { version = "=0.7.0-a.1", path = "../../aj" }
 serde = { version = "1.0.64", features = ["derive"] }
 actix-rt = "2.2"
 chrono = { version = "0.4.34" }
