--- examples/retry/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "retry"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-aj = { version = "=0.6.4" }
-serde = { version = "1.0.64", features = ["derive"] }
-actix-rt = "2.2"
-chrono = { version = "0.4.34" }
