--- examples/simple_job/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "simple_job"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-aj = { version = "=0.6.4" }
-serde = { version = "1.0.64", features = ["derive"] }
-actix-rt = "2.2"
