+++ examples/tokio/Cargo.toml
@@ -5,5 +5,5 @@ edition = "2021"
 
 [dependencies]
 tokio = { version = "^1.40.0", features = ["full"] }
-aj = { version = "0.6.6" }
+aj = { version = "=0.7.0-a.1", path = "../../aj" }
 serde = { version = "1.0.64", features = ["derive"] }
